refactor: rename NavigateToExternalUrl module as smart-urls
